Move Deserializer into its own subdirectory
